reorganized repo and a weird bug with corpses being hit by spikeball
