+++ spikeball_trap/chain_half.png
