+++ spikeball_trap/spikeball.png
